requirements.txt
flask==2.0.1
flask-cors==3.0.10
mysql-connector-python==8.0.27
nltk==3.6.5
python-dotenv==0.19.1
gunicorn==20.1.0